+++ mulog-elasticsearch/src/com/brunobonacci/mulog/publishers/elasticsearch.clj
@@ -52,11 +52,23 @@
 
 
 
-(defmethod index-name :data-stream [[_ v]] {:op :create :index* (constantly v)})
+;;
+;; Although when using data-stream the client should
+;; use the `create` action, we have experienced issues with it
+;; (see: https://github.com/BrunoBonacci/mulog/issues/92).
+;;
+;; The main problem is that the `create` action is not idempotent,
+;; thus in case of partial failure it won't accept an event with an
+;; `_id` which already exists.  Using the `index` operation solves
+;; this problem.
+;;
+(defmethod index-name :data-stream [[_ v]]
+  {:op :index :index* (constantly v)})
 
 
 
-(defmethod index-name :default [_] (index-name [:index-pattern "'mulog-'yyyy.MM.dd"]))
+(defmethod index-name :default [_]
+  (index-name [:index-pattern "'mulog-'yyyy.MM.dd"]))
 
 
 
@@ -121,12 +133,12 @@
           (update :body json/from-json))]
     ;; ELS BulkAPI respond with HTTP 200 even if there are failing
     ;; items. See #79
+    ;; see: https://www.elastic.co/guide/en/elasticsearch/reference/current/docs-bulk.html
     (if (-> response :body :errors)
       (throw (ex-info "Elasticsearch Bulk API reported errors"
-               {:errors (filter #(>= (-> % :index :status) 400)
+               {:errors (remove #(< (-> % :index :status (or 999)) 400)
                           (-> response :body :items))}))
       response)))
-
 
 
 
@@ -157,12 +169,13 @@
 (comment
 
 
-  (apply-defaults {:url "http://localhost:9200" :data-stream "mulog-stream"})
+  (apply-defaults {:url "http://localhost:9200" :data-stream "mulog-stream" :els-version :v7.x})
 
 
   (prepare-records
     (apply-defaults
       {:url "http://localhost:9200/_bulk"
+       :els-version :v7.x
        :name-mangling true})
     [{:mulog/timestamp (System/currentTimeMillis) :event-name :hello :k 1}
      {:mulog/timestamp (System/currentTimeMillis) :event-name :hello :k nil}])
@@ -223,8 +236,11 @@
   {;; :url endpoint for Elasticsearch
    ;; :url "http://localhost:9200/" ;; REQUIRED
    :max-items     5000
+
    :publish-delay 5000
+
    :name-mangling true
+
    ;; Choose between `:index-pattern` or `:data-stream`, the default is `:index-pattern`
    ;; The pattern uses the Java DateTimeFormatter format:
    ;; see: https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/time/format/DateTimeFormatter.html
@@ -232,9 +248,12 @@
    ;;
    ;; data streams are available since Elasticsearch 7.9
    ;; :data-stream   "mulog-stream"
+
    ;; extra http options
    :http-opts {}
-   :els-version   :auto   ;; one of: `:v6.x`, `:v7.x`, `:auto`
+
+   :els-version   :auto   ;; one of: `:v6.x`, `:v7.x`, `:v8.x`, `:auto`
+
    ;; function to transform records
    :transform     identity})
 
@@ -249,7 +268,7 @@
     ;; autodetect version when set to `:auto`
     (update $ :els-version
       (fn [v] (if (= v :auto)
-                (or (detect-els-version url (:http-opts $)) :v7.x)
+                (or (detect-els-version url (:http-opts $)) :v8.x)
                 v)))))